+++ src/languages/twig.js
@@ -2,6 +2,7 @@
 Language: Twig
 Requires: xml.js
 Author: Luke Holder <lukemh@gmail.com>
+Contributors: Florian Förster <foerster.florianjr@gmail.com>
 Description: Twig is a templating language for PHP
 Website: https://twig.symfony.com
 Category: template
@@ -13,7 +14,7 @@ export default function(hljs) {
     begin: '\\(', end: '\\)'
   };
 
-  var FUNCTION_NAMES = 'attribute block constant cycle date dump include ' +
+  var FUNCTION_NAMES = 'attribute block constant country_timezones cycle date dump html_classes include ' +
                   'max min parent random range source template_from_string';
 
   var FUNCTIONS = {
@@ -28,17 +29,17 @@ export default function(hljs) {
   var FILTER = {
     begin: /\|[A-Za-z_]+:?/,
     keywords:
-      'abs batch capitalize column convert_encoding date date_modify default ' +
-      'escape filter first format inky_to_html inline_css join json_encode keys last ' +
-      'length lower map markdown merge nl2br number_format raw reduce replace ' +
-      'reverse round slice sort spaceless split striptags title trim upper url_encode',
+      'abs batch capitalize column convert_encoding country_name currency_name currency_symbol data_uri date date_modify default ' +
+      'escape filter first format format_currency format_date format_datetime format_number format_time html_to_markdown inky_to_html inline_css join json_encode keys language_name last ' +
+      'length locale_name lower map markdown_to_html merge nl2br number_format raw reduce replace ' +
+      'reverse round slice sort spaceless split striptags timezone_name title trim u upper url_encode',
     contains: [
       FUNCTIONS
     ]
   };
 
-  var TAGS = 'apply autoescape block deprecated do embed extends filter flush for from ' +
-    'if import include macro sandbox set use verbatim with';
+  var TAGS = 'apply autoescape block deprecated do embed extends flush for from ' +
+    'if import in include is macro sandbox set use verbatim with';
 
   TAGS = TAGS + ' ' + TAGS.split(' ').map(function(t){return 'end' + t}).join(' ');